refactor ke dto
